feat: separando a lógica de estado e a submissão do componente de formulário
